requirements.txt
PyYAML==5.4.1
Jinja2==3.1.2
pdfkit==1.0.0